backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-dotenv==1.0.1
motor==3.6.0
pymongo==4.9.0
pydantic==2.9.2
bcrypt==4.2.0
PyJWT==2.9.0
mistralai==1.1.0
python-docx==1.1.2
pdfminer.six==20231228
Jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.2
slowapi==0.1.9
pytest==8.3.3
pytest-asyncio==0.24.0
sentence-transformers==5.2.0
faiss-cpu==1.13.2
numpy==2.2.1
